+++ core-base-fe/src/index.ts
@@ -15,13 +15,17 @@ import { MoneyField } from './react-admin/fields/MoneyField'
 import { StatusInput } from './react-admin/inputs/StatusInput'
 import { MoneyInput } from './react-admin/inputs/MoneyInput'
 import { generateTimeZoneOptions, TimezoneSelectInput } from './react-admin/inputs/TimezoneSelectInput'
+import { clickToCopy } from './react-admin/fields/clickToCopy'
+import './react-admin/fields/clickToCopy.css'
+import { ToggleButtonInput } from './react-admin/inputs/ToggleButtonInput'
+import { ExternalLink } from './components/ExternalLink'
 
 
 /**
  * function utilities
  * @public
  */
-export { timeAgo, generateTimeZoneOptions }
+export { timeAgo, generateTimeZoneOptions, clickToCopy }
 
 /**
  * React Admin Fields
@@ -33,4 +37,10 @@ export { DateField, TimeAgoField, PolymorphicReferenceField, MoneyField }
  * React Admin Inputs
  * @public
  */
-export { StatusInput, MoneyInput, TimezoneSelectInput }
+export { StatusInput, MoneyInput, TimezoneSelectInput, ToggleButtonInput }
+
+/**
+ * MUI Components
+ * @public
+ */
+export { ExternalLink }