+++ core-base-fe/src/react-admin/fields/clickToCopy.ts
@@ -1,23 +1,8 @@
-// TODO Enable this function - should have a default notification behavior that can be disabled or delegated to the calling app
-
-// const clickToCopy = {
-//   onClick: (element: React.MouseEvent<HTMLButtonElement>) => {
-//     navigator.clipboard.writeText(element.currentTarget.innerText)
-//     notify(`Copied`, {
-//       type: 'success',
-//       anchorOrigin: { vertical: 'top', horizontal: 'center' },
-//     })
-//   },
-//   sx: { cursor: 'copy' },
-// }
-
 import './clickToCopy.css'
 
 export const clickToCopy = {
   onClick: (element: React.MouseEvent<HTMLButtonElement>) => {
     navigator.clipboard.writeText(element.currentTarget.innerText)
-    // TODO
-    // Change the mouse cursor for .5s to 'progress' to indicate that the text has been copied
   },
   className: 'copyable-content',
 }