+++ core-base-fe/stories/TimezoneSelectInput.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { withFormContext  } from './RAStorybookDecorators'
+import { withFormContext  } from '../../stories/RAStorybookDecorators'
 import { TimezoneSelectInput } from '../src/react-admin/inputs/TimezoneSelectInput'
 import { required } from 'react-admin';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: 'CORE/React Admin/Inputs/TimezoneSelectInput',
+  title: 'React Admin/Inputs/TimezoneSelectInput',
   component: TimezoneSelectInput,
   decorators: [withFormContext],
   tags: ['autodocs'],  // https://storybook.js.org/docs/react/writing-docs/autodocs,