+++ core-webhooks-fe/src/core-webhooks/index.ts
@@ -1,4 +1,3 @@
 // WebhookEvent
 
-// import { Charts } from './views'
 export { Webhooks } from './resources/webhooks'