+++ test-core-fe/src/App.tsx
@@ -5,13 +5,9 @@ import { Admin, Resource } from 'react-admin'
 
 // Source Code import
 import { Comments } from '@moonlight-labs/core-comments-fe'
-
 import { Jobs } from '@moonlight-labs/core-jobs-fe'
-// import { CoreAccounting } from '@moonlight-labs/core-accounting-fe'
-
-// import { CoreVersions } from '@moonlight-labs/core-versions-fe'
 import { Webhooks } from '@moonlight-labs/core-webhooks-fe'
-
+import { Notifications } from '@moonlight-labs/core-notifications-fe'
 // Wireframed Version of the resource (Local)
 // Uncomment this to swap in the earlier version
 // import { Jobs } from './jobs/resource'
@@ -98,7 +94,15 @@ function AdminApp() {
         name="Webhook"
         icon={Webhooks.Icon}
         list={Webhooks.List}
-        edit={Webhooks.Show}
+        show={Webhooks.Show}
+        recordRepresentation={Webhooks.recordRepresentation}
+      />
+
+      <Resource
+        name="Notification"
+        icon={Notifications.Icon}
+        list={Notifications.List}
+        edit={Notifications.Edit}
       />
     </Admin>
   )