+++ core-base-fe/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @moonlight-labs/core-base-fe
 
+## 0.2.7
+
+### Patch Changes
+
+- Base enhancements, Webhooks improvements
+
+## 0.2.6
+
+### Patch Changes
+
+- TimezonSelectInput and Versions improvements
+
 ## 0.2.5
 
 ### Patch Changes