+++ core-comments-fe/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @moonlight-labs/core-comments-fe
 
+## 0.2.7
+
+### Patch Changes
+
+- Base enhancements, Webhooks improvements
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.7
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @moonlight-labs/core-base-fe@0.2.6
+
 ## 0.2.5
 
 ### Patch Changes